+++ web_backup_2026-04-27/runtime-config.js
@@ -1,6 +1,0 @@
-window.VIBE_CONFIG = Object.assign(
-  {
-    API_BASE: "",
-  },
-  window.VIBE_CONFIG || {}
-);